Issue #682 Strain Lineage tree - fix remarks
