+++ server/catgenome/src/main/java/com/epam/catgenome/dao/lineage/LineageTreeNodeDao.java
@@ -43,8 +43,8 @@ import java.sql.Timestamp;
 import java.util.List;
 import java.util.Set;
 
-import static com.epam.catgenome.manager.lineage.LineageTreeManager.serializeAttributes;
-import static com.epam.catgenome.manager.lineage.LineageTreeManager.parseAttributes;
+import static com.epam.catgenome.util.Utils.parseAttributes;
+import static com.epam.catgenome.util.Utils.serializeAttributes;
 
 @Getter
 @Setter
