+++ server/catgenome/src/main/resources/catgenome-messages.properties
@@ -291,3 +291,6 @@ error.lineage.node.name.required=Incorrect file format. Node name should be defi
 error.lineage.node.not.found=Incorrect edges file format. Node ''{0}'' not found
 error.lineage.not.unique.node.name=Incorrect nodes file format. Node ''{0}'' not unique
 error.lineage.incorrect.edge=Incorrect edges file format. Node ''{0}'' can not be connected to itself
+
+#attributes
+error.attribute.invalid.value=Invalid attribute value: ''{0}''
